backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.0
groq==0.11.0
supabase==2.9.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
stripe==10.0.0
textblob==0.18.0
numpy==1.26.0
Pillow==10.4.0
python-multipart==0.0.12
